Extract fetch roster code to use case
